refactor: перенести library-query-parser из pages в components
